ZotCore
- исправлен random tick block
- теперь id блока хранится в качестве int, а не byte

+++ src/main/java/cn/nukkit/level/format/ChunkSection.java
@@ -36,7 +36,7 @@ public interface ChunkSection {
 
     void setBlockLight(int x, int y, int z, int level);
 
-    byte[] getIdArray();
+    int[] getIdArray();
 
     byte[] getDataArray();
